change: - cleanup of styles - cleanup on menu and route names. - update select component to consume value. - Vote And Dao Folder restructure.
