+++ packages/boba/gateway/src/components/select/Select.js
@@ -124,6 +124,7 @@ function Select({
     ].join(' ')}>
       {label && <Box className={styles.label}>{label}</Box>}
       <ReactSelect
+        value={value}
         onChange={onSelect}
         isMulti={isMulti}
         options={options}
