+++ packages/boba/gateway/src/containers/VoteAndDao/Dao/Dao.styles.js
@@ -43,19 +43,12 @@ export const DaoWalletContainer = styled(Box)(({ theme }) => ({
   padding: '0px 20px',
   minHeight: '700px',
   width: '30%',
+  gap: '10px',
   borderRadius: theme.palette.primary.borderRadius,
   background: theme.palette.background.secondary,
   [theme.breakpoints.down('sm')]: {
     width: '100%',
   },
-}));
-
-export const DaoWalletAction = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  justifyContent: 'space-around',
-  width: '100%',
-  margin: '10px auto',
-  gap: '10px',
 }));
 
 export const DaoProposalContainer = styled(Box)(({ theme }) => ({
