+++ packages/boba/gateway/src/containers/VoteAndDao/Vote/Pools/pools.tableHeads.js
