--- packages/boba/gateway/src/containers/dao/Dao.module.scss
@@ -1,84 +1,0 @@
-@import 'index.scss';
-
-.container {
-  display: flex;
-  flex-direction: column;
-  padding: 10px;
-  padding-top: 0px;
-
-  .header {
-    h2 {
-      margin-bottom: 0;
-      padding-left: 10px;
-      color: $white;
-    }
-
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  .content {
-
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: flex-start;
-    width: 100%;
-    border-radius: 4px;
-    gap: 10px;
-
-    @media only screen and (max-width: 900px) {
-      flex-direction: column;
-    }
-
-    .topRow {
-      display: flex;
-      flex-direction: row;
-      @media only screen and (max-width: 550px) {
-        flex-direction: column;
-      }
-    }
-
-    .transferContainer,
-    .delegateContainer {
-      display: flex;
-      flex-direction: column;
-      justify-content: space-around;
-      align-items: flex-start;
-      width: 250px;
-      @media only screen and (max-width: 900px) {
-        width: 100%;
-        min-width: unset;
-        min-height: unset;
-      }
-      .info {
-        .title {
-          font-size: 1.4em;
-          color: $white;
-        }
-        .subTitle {
-          font-size: 1.0em;
-          color: blue;
-          margin: 10px;
-        }
-        .helpText {
-          font-size: 0.8em;
-          margin: 10px;
-          margin-bottom: 10px;
-          //opacity: 0.8;
-        }
-        .helpTextLight {
-          font-size: 0.7em;
-          margin: 10px;
-          opacity: 0.6;
-          text-align: left; 
-          line-height: 1.0em; 
-          margin-top: 20px;
-          padding: 20px;
-        }
-      }
-    }
-  }
-}
