--- packages/boba/gateway/src/containers/dao/Pools/poolList.style.js
@@ -1,11 +1,0 @@
-import { styled } from '@mui/material/styles'
-import { Box } from "@mui/material"
-
-export const ListItemContent = styled(Box)(({ theme }) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '20px',
-  width: '100%',
-  padding: '10px',
-  borderBottom: theme.palette.primary.borderBottom
-}))
