+++ packages/boba/gateway/src/layout/index.js
@@ -38,10 +38,11 @@ import Bridge from 'containers/bridge/Bridge'
 import MonsterWrapper from 'containers/monster/MonsterWrapper'
 import Lock from 'containers/veboba/Lock'
 import FarmWrapper from 'containers/farm/FarmWrapper'
-import Dao from 'containers/dao/Dao'
+// import Dao from 'containers/VoteAndDao/Dao/Dao'
 import SaveWrapper from 'containers/save/SaveWrapper'
 import Projects from 'containers/ecosystem/Projects'
 import { ROUTES_PATH } from 'util/constant'
+import VoteAndDao from 'containers/VoteAndDao'
 
 function App() {
